renamme dummy folder name to buttons
