+++ web/src/components/buttons/RemediationButton/graphql/remediateResource.generated.ts
