+++ web/src/components/buttons/RemediationButton/graphql/remediateResource.graphql
