+++ web/src/components/forms/AnalyticsConsentForm/AnalyticsConsentForm.tsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import { Formik } from 'formik';
 import * as Yup from 'yup';
-import SubmitButton from 'Components/Buttons2/SubmitButton';
+import SubmitButton from 'Components/buttons/SubmitButton';
 import ErrorReportingSection from './ErrorReportingSection';
 
 interface AnalyticsConsentFormValues {
