+++ web/src/components/forms/ForgotPasswordConfirmForm/ForgotPasswordConfirmForm.tsx
@@ -20,7 +20,7 @@ import React from 'react';
 import { Field, Formik } from 'formik';
 import * as Yup from 'yup';
 import { Alert, Box, useSnackbar } from 'pouncejs';
-import SubmitButton from 'Components/Buttons2/SubmitButton';
+import SubmitButton from 'Components/buttons/SubmitButton';
 import FormikTextInput from 'Components/fields/TextInput';
 import useRouter from 'Hooks/useRouter';
 import useAuth from 'Hooks/useAuth';
