+++ web/src/components/forms/TotpForm/TotpForm.tsx
@@ -22,7 +22,7 @@ import QRCode from 'qrcode.react';
 import * as React from 'react';
 import * as Yup from 'yup';
 import { formatSecretCode } from 'Helpers/utils';
-import SubmitButton from 'Components/Buttons2/SubmitButton';
+import SubmitButton from 'Components/buttons/SubmitButton';
 import FormikTextInput from 'Components/fields/TextInput';
 import useAuth from 'Hooks/useAuth';
