+++ web/src/pages/PolicyDetails/columns.tsx
@@ -22,8 +22,8 @@ import React from 'react';
 import { Text, TableProps, Tooltip, Label } from 'pouncejs';
 import { ComplianceItem, ComplianceStatusEnum } from 'Generated/schema';
 import { capitalize, formatDatetime } from 'Helpers/utils';
-import SuppressButton from 'Components/Buttons2/SuppressButton';
-import RemediationButton from 'Components/Buttons2/RemediationButton';
+import SuppressButton from 'Components/buttons/SuppressButton';
+import RemediationButton from 'Components/buttons/RemediationButton';
 
 // The columns that the associated table will show
 const columns = [
